requirements.txt
numpy>=1.24
opencv-python>=4.8
Pillow>=9.5
PyYAML>=6.0
scipy>=1.10
scikit-image>=0.21
matplotlib>=3.7
pandas>=1.5
seaborn>=0.12
psutil>=5.9
requests>=2.31
tqdm>=4.66
tensorboard>=2.14
thop>=0.1.1
pycocotools>=2.0.7
rawpy>=0.18
torch>=2.0
torchvision>=0.15
ultralytics>=8.2
